requirements.txt
sqlmodel
fastapi[standard]
python-dotenv
psycopg2-binary
alembic